pc/src/requirements.txt
proxy-tools==0.1.0
pycparser==2.21
pythonnet==2.5.2
pywebview==3.6.2
